Changes how DataConfig is initialized (NOTES).
We had a separate function for creating a config from a dict, which took an already-created instance of the DataConfig and then overwrote default values. I modified this so that you can just pass in a dict using **dict and then the values are either set to the default or taken from the dict, if available.

This means that anyone who uses config files or other tools to create DataConfig objects just needs to convert those files or tools to dictionaries. More flexible, right?