Removes flexible kernel sizes for flat and dilation net (NOTE).
We may still want this, but it's not good as implemented, where different architectures have different expectations and the expectations are not clearly defined. I'm turning it off so that it works with the standard kernel_size format, and we can reimplement when we know how we'd like to have that in the future.